# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"Bearer Authentication":[]}],"paths":{"/api/v1/focus/sessions/{sessionUuid}/end":{"post":{"tags":["focus-session-controller"],"operationId":"completeSession","parameters":[{"name":"sessionUuid","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompleteFocusSessionRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CompleteFocusSessionResponse"}}}}}}},"/api/v1/focus/sessions/start":{"post":{"tags":["focus-session-controller"],"operationId":"startSession","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StartFocusSessionRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StartFocusSessionResponse"}}}}}}},"/api/v1/auth/refresh":{"post":{"tags":["auth-controller"],"operationId":"refresh","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}}}}},"/api/v1/users/me":{"get":{"tags":["user-controller"],"operationId":"me","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/species":{"get":{"tags":["species-controller"],"operationId":"listSpecies","parameters":[{"name":"includeDisabled","in":"query","required":false,"schema":{"type":"boolean","default":false}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SpeciesListResponse"}}}}}}},"/api/v1/inventory/trees":{"get":{"tags":["tree-inventory-controller"],"operationId":"listInventory","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TreeInventoryListResponse"}}}}}}}},"components":{"schemas":{"CompleteFocusSessionRequest":{"type":"object","properties":{"clientEndTime":{"type":"string","format":"date-time"},"state":{"type":"string","enum":["CREATED","SUCCESS","INTERRUPTED"]}},"required":["state"]},"CompleteFocusSessionResponse":{"type":"object","properties":{"state":{"type":"string","enum":["CREATED","SUCCESS","INTERRUPTED"]},"validatedMinutes":{"type":"integer","format":"int32"},"softCurrencyAwarded":{"type":"integer","format":"int32"},"wallet":{"$ref":"#/components/schemas/WalletSummary"}}},"WalletSummary":{"type":"object","properties":{"softCurrency":{"type":"integer","format":"int32"},"hardCurrency":{"type":"integer","format":"int32"}}},"StartFocusSessionRequest":{"type":"object","properties":{"sessionUuid":{"type":"string","format":"uuid"},"speciesCode":{"type":"string","maxLength":64,"minLength":0,"pattern":"^[A-Za-z0-9_]+$"},"clientStartTime":{"type":"string","format":"date-time"},"plannedMinutes":{"type":"integer","format":"int32","maximum":120,"minimum":5},"tag":{"type":"string","maxLength":20,"minLength":0,"pattern":"^[A-Za-z0-9 ]+$"}},"required":["clientStartTime","sessionUuid"]},"StartFocusSessionResponse":{"type":"object","properties":{"sessionUuid":{"type":"string","format":"uuid"},"serverStartTime":{"type":"string","format":"date-time"},"state":{"type":"string","enum":["CREATED","SUCCESS","INTERRUPTED"]}}},"RefreshRequest":{"type":"object","properties":{"refreshToken":{"type":"string"},"clientId":{"type":"string"}}},"TokenResponse":{"type":"object","properties":{"accessToken":{"type":"string"},"expiresIn":{"type":"integer","format":"int64"},"refreshToken":{"type":"string"},"tokenType":{"type":"string"}}},"LoginRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"clientId":{"type":"string"}}},"SpeciesItem":{"type":"object","properties":{"uuid":{"type":"string","format":"uuid"},"code":{"type":"string"},"name":{"type":"string"},"isPremium":{"type":"boolean"},"price":{"type":"integer","format":"int32"},"isEnabled":{"type":"boolean"},"sortOrder":{"type":"integer","format":"int32"}}},"SpeciesListResponse":{"type":"object","properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/SpeciesItem"}}}},"TreeInventoryItem":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"speciesUuid":{"type":"string","format":"uuid"},"speciesCode":{"type":"string"},"speciesName":{"type":"string"},"placed":{"type":"boolean"},"cellX":{"type":"integer","format":"int32"},"cellY":{"type":"integer","format":"int32"}}},"TreeInventoryListResponse":{"type":"object","properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/TreeInventoryItem"}}}}},"securitySchemes":{"Bearer Authentication":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}